pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dotX12 <dev@shitposting.team>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
SQLAlchemy = "^1.4.42"
pydantic = "^1.10.2"
asyncpg = "^0.26.0"
uvicorn = "^0.18.3"
alembic = "^1.8.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"
cashews = "^4.7.1"
psycopg2-binary = "^2.9.4"
redis = "^4.3.4"
uuid7 = "^0.1.0"
sqlalchemy2-stubs = "^0.0.2-alpha.29"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
